MFB: fix #35781 (stream_filter_append() can cause segfault)

+++ main/streams/plain_wrapper.c
@@ -16,7 +16,7 @@
    +----------------------------------------------------------------------+
  */
 
-/* $Id: plain_wrapper.c,v 1.54 2005/11/17 14:19:06 tony2001 Exp $ */
+/* $Id: plain_wrapper.c,v 1.55 2005/12/23 15:05:42 tony2001 Exp $ */
 
 #include "php.h"
 #include "php_globals.h"
@@ -316,7 +316,7 @@ static size_t php_stdiop_read(php_stream *stream, char *buf, size_t count TSRMLS
 	if (data->fd >= 0) {
 		ret = read(data->fd, buf, count);
 		
-		stream->eof = (ret == 0 || (ret == -1 && errno != EWOULDBLOCK));
+		stream->eof = (ret == 0 || (ret == (size_t)-1 && errno != EWOULDBLOCK));
 				
 	} else {
 #if HAVE_FLUSHIO
